requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=9.0.0
scipy>=1.7.0
matplotlib>=3.5.0
